feat: Context 관심사 분리를 통한 App 컴포넌트 최적화
